+++ R/decode.R
@@ -237,18 +237,19 @@ setAs(from = "corpus", to = "AnnotatedPlainTextDocument", def = function(from){
 #' }
 #' }
 #' @rdname decode
+#' @importFrom cli cli_progress_step
 setMethod("decode", "corpus", function(.Object, to = c("data.table", "Annotation"), p_attributes = NULL, s_attributes = NULL, decode = TRUE, verbose = TRUE){
   if (to == "data.table"){
 
     if (is.null(p_attributes)) p_attributes <- p_attributes(.Object)
     if (!all(p_attributes %in% p_attributes(.Object)))
       stop("Not all p_attributes provided are available.")
-    
 
     p_attribute_list <- lapply(
       setNames(p_attributes, p_attributes),
       function(p_attr){
-        if (verbose) message("decoding p-attribute:", p_attr)
+        if (verbose)
+          cli_progress_step(sprintf("decoding p-attribute: %s", p_attr))
         get_token_stream(.Object, p_attribute = p_attr, decode = decode)
       }
     )
@@ -265,7 +266,8 @@ setMethod("decode", "corpus", function(.Object, to = c("data.table", "Annotation
       s_attribute_list <- lapply(
         setNames(s_attributes, s_attributes),
         function(s_attr){
-          if (verbose) message("decoding s-attribute:", s_attr)
+          if (verbose)
+            cli_progress_step(sprintf("decoding s-attribute: %s", s_attr))
           struc <- cl_cpos2struc(
             corpus = .Object@corpus, registry = .Object@registry_dir,
             s_attribute = s_attr, cpos = 0L:max_cpos
@@ -286,7 +288,7 @@ setMethod("decode", "corpus", function(.Object, to = c("data.table", "Annotation
       s_attribute_list <- list()
     }
 
-    message("assembling data.table")
+    if (verbose) cli_progress_step("assembling data.table")
     combined_list <- c(
       list(cpos = 0L:max_cpos),
       p_attribute_list,